test: relocate VLM e2e tests into hf_transformer_vlm
