+++ tests/functional_tests/hf_transformer/test_hf_transformer.py
@@ -14,27 +14,12 @@
 
 import shutil
 
-import pytest
-
 from tests.utils.test_utils import run_test_script
-
-try:
-    import qwen_vl_utils  # noqa: F401
-
-    _has_qwen_vl_utils = True
-except ImportError:
-    _has_qwen_vl_utils = False
-
-skip_if_no_qwen_vl_utils = pytest.mark.skipif(
-    not _has_qwen_vl_utils,
-    reason="qwen_vl_utils is not installed (requires the vlm-media extra)",
-)
 
 TEST_FOLDER = "hf_transformer"
 HF_TRANSFORMER_FUSED_CE_SFT_FILENAME = "L2_HF_Transformer_Fused_CE_SFT.sh"
 HF_TRANSFORMER_PACKED_SEQUENCE_FILENAME = "L2_HF_Transformer_Packed_Sequence.sh"
 HF_TRANSFORMER_LLM_META_FILENAME = "L2_HF_Transformer_LLM_Meta.sh"
-HF_TRANSFORMER_VLM_META_FILENAME = "L2_HF_Transformer_VLM_Meta.sh"
 
 
 class TestHFTransformer:
@@ -55,14 +40,6 @@ class TestHFTransformer:
     def test_hf_transformer_llm_meta(self):
         try:
             run_test_script(TEST_FOLDER, HF_TRANSFORMER_LLM_META_FILENAME)
-        finally:
-            # remove the checkpoint directory
-            shutil.rmtree("checkpoints/", ignore_errors=True)
-
-    @skip_if_no_qwen_vl_utils
-    def test_hf_transformer_vlm_meta(self):
-        try:
-            run_test_script(TEST_FOLDER, HF_TRANSFORMER_VLM_META_FILENAME)
         finally:
             # remove the checkpoint directory
             shutil.rmtree("checkpoints/", ignore_errors=True)
