+++ tests/functional_tests/hf_transformer_vlm/L2_HF_PEFT_VLM_FSDP2_Checkpoint.sh
