+++ tests/functional_tests/hf_transformer_vlm/L2_HF_Transformer_VLM_Meta.sh
