+++ tests/functional_tests/hf_transformer_vlm/test_hf_transformer_vlm.py
@@ -26,6 +26,8 @@ HF_TRANSFORMER_VLM_PEFT_FILENAME = "L2_HF_Transformer_VLM_PEFT.sh"
 HF_TRANSFORMER_VLM_SFT_FILENAME = "L2_HF_Transformer_VLM_SFT.sh"
 HF_TRANSFORMER_VLM_SFT_MegatronFSDP_FILENAME = "L2_HF_Transformer_VLM_SFT_MegatronFSDP.sh"
 HF_TRANSFORMER_VLM_GEMMA4_JOINT_DRAFTER_FILENAME = "L2_HF_Transformer_VLM_Gemma4_Joint_Drafter.sh"
+HF_PEFT_VLM_FSDP2_CHECKPOINT_FILENAME = "L2_HF_PEFT_VLM_FSDP2_Checkpoint.sh"
+HF_TRANSFORMER_VLM_META_FILENAME = "L2_HF_Transformer_VLM_Meta.sh"
 
 
 class TestHFTransformerVLM:
@@ -67,3 +69,17 @@ class TestHFTransformerVLM:
             run_test_script(TEST_FOLDER, HF_TRANSFORMER_VLM_GEMMA4_JOINT_DRAFTER_FILENAME)
         finally:
             shutil.rmtree("vlm_checkpoints/", ignore_errors=True)
+
+    def test_hf_peft_vlm_fsdp2_checkpoint(self):
+        try:
+            run_test_script(TEST_FOLDER, HF_PEFT_VLM_FSDP2_CHECKPOINT_FILENAME)
+        finally:
+            # remove the checkpoint directory
+            shutil.rmtree("checkpoints/", ignore_errors=True)
+
+    def test_hf_transformer_vlm_meta(self):
+        try:
+            run_test_script(TEST_FOLDER, HF_TRANSFORMER_VLM_META_FILENAME)
+        finally:
+            # remove the checkpoint directory
+            shutil.rmtree("checkpoints/", ignore_errors=True)
